Fix 32-bit large-file digest truncation
Large-file builds use a 64-bit off_t on 32-bit systems, but get_hash_from_fd2 accepted the byte count as size_t. Files at or above 4 GiB were therefore hashed modulo 2^32 bytes, allowing trust creation and SHA or IMA fallback verification to agree on a truncated digest.

Carry file sizes as off_t through the hashing API and the Debian MD5 backend. Preserve the mmap path for representable sizes, and use offset-safe pread chunks with incremental EVP hashing when the size exceeds SIZE_MAX.

Add an i686-gated LMDB regression that trusts an unchanged 4 GiB file, mutates its final byte, and requires SHA verification to reject it. The RPM and Debian build/test matrices pass, and a focused ELF32 run confirmed that changing the last byte changes the digest.

+++ src/library/file.h
@@ -106,7 +106,7 @@ char *get_file_type_from_fd(int fd, const struct file_info *i, const char *path,
 	__attr_access ((__write_only__, 5, 4)) __attr_fd_arg_read (1);
 char *bytes2hex(char *final, const unsigned char *buf, unsigned int size)
 	 __attr_access ((__read_only__, 2, 3));
-char *get_hash_from_fd2(int fd, size_t size, file_hash_alg_t alg)
+char *get_hash_from_fd2(int fd, off_t size, file_hash_alg_t alg)
 	__attr_dealloc_free __attr_fd_arg_read (1);
 int get_ima_hash(int fd, file_hash_alg_t *alg, char *sha)
 	__attr_fd_arg (1);
